move generate_source out of helpers
