+++ src/helpers/mod.ts
@@ -1,6 +1,5 @@
 export * from "./ask_trivia.ts";
 export * from "./combine_answers.ts";
-export * from "./generate_source.ts";
 export * from "./get_user_input.ts";
 export * from "./logger.ts";
 export * from "./pick_trivia.ts";
