+++ src/trivia/generate_source.ts
