+++ src/trivia/television/the_clone_wars.ts
@@ -1,5 +1,5 @@
 import { TRIVIA } from "types";
-import { generateTVSource } from "helpers";
+import { generateTVSource } from "../generate_source.ts";
 
 const genSource = (episode: string) =>
   generateTVSource("The Clone Wars", episode);
